Add support for @time timestamps in ZNC playback buffer
This adds proper display support for the playback buffer timestamps of
ZNC 1.5+ in combination with mIRC 7.33+, which will use the new
server-time and znc.in/server-time[-iso] IRCv3 extension instead of a
simple timestamp prepended to the saved text. This was causing nbs-irc
to just display the current time on connect instead of the proper past
time of the messages in the playback buffer.